build: Restructure inheritance
We brought up a dedicated tool (docker-recipe) for
centralizing configurations
and solving inheritance.
This is a huge overhead as we only offer PHP-container so far.
Therefor the inheritance is now solved via symlinks
and file prefixes (to compile in the correct order).
This makes the project standalone again
so people can compile easier.